+++ android/src/main/java/com/github/rmtmckenzie/native_device_orientation/IOrientationListener.java
@@ -1,4 +1,4 @@
-package com.github.rmtmckenzie.nativedeviceorientation;
+package com.github.rmtmckenzie.native_device_orientation;
 
 public interface IOrientationListener {